Created an upload_handler service.
