+++ dropzonejs.services.yml
@@ -2,3 +2,6 @@ services:
   dropzonejs.upload_save:
     class: Drupal\dropzonejs\DropzoneJsUploadSave
     arguments: ['@entity.manager', '@file.mime_type.guesser', '@file_system', '@logger.factory', '@renderer', '@config.factory']
+  dropzonejs.upload_handler:
+    class: Drupal\dropzonejs\UploadHandler
+    arguments: ['@request_stack', '@config.factory', '@transliteration']
