+++ src/UploadHandlerInterface.php
@@ -1,0 +1,40 @@
+<?php
+
+/**
+ * @file
+ * Contains \Drupal\dropzonejs\UploadHandlerInterface.
+ */
+
+namespace Drupal\dropzonejs;
+
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+
+interface UploadHandlerInterface {
+
+  /**
+   * Reads, checks and return filename of a file being uploaded.
+   *
+   * @param \Symfony\Component\HttpFoundation\File\UploadedFile $file
+   *   An instance of UploadedFile.
+   *
+   * @return string
+   *   The sanitized filename.
+   *
+   * @throws \Drupal\dropzonejs\UploadException
+   */
+  public function getFilename(UploadedFile $file);
+
+  /**
+   * Handles an uploaded file.
+   *
+   * @param \Symfony\Component\HttpFoundation\File\UploadedFile $file
+   *   The uploaded file.
+   *
+   * @return string
+   *   URI of the uploaded file.
+   *
+   * @throws \Drupal\dropzonejs\UploadException
+   */
+  public function handleUpload(UploadedFile $file);
+
+}
